+++ tests/test_models.py
@@ -7,18 +7,29 @@ from apisports_football.Models.Fixtures.Events import Events
 from apisports_football.Models.Fixtures.Fixtures import Fixtures
 from apisports_football.Models.Fixtures.Lineups import Lineups
 from apisports_football.Models.Fixtures.Players import Players as FixturesPlayers
-from apisports_football.Models.Fixtures.Statistics import Statistics
+from apisports_football.Models.Fixtures.Statistics import Statistics as FixturesStatistics
 from apisports_football.Models.Fixtures.Rounds import Rounds
 from apisports_football.Models.Injuries.Injuries import Injuries
 from apisports_football.Models.Leagues.Leagues import Leagues
 from apisports_football.Models.Leagues.Seasons import Seasons
 from apisports_football.Models.Odds.Odds import Odds
 from apisports_football.Models.Odds.Bookmakers import Bookmakers
+from apisports_football.Models.Odds.Live import Live
 from apisports_football.Models.Odds.LiveBets import LiveBets
 from apisports_football.Models.Odds.Mapping import Mapping
 from apisports_football.Models.Players.Players import Players
 from apisports_football.Models.Players.Profiles import Profiles
-from apisports_football.Models.Players.Teams import Teams
+from apisports_football.Models.Players.Squads import Squads
+from apisports_football.Models.Players.Teams import Teams as PlayersTeams
+from apisports_football.Models.Predictions.Predictions import Predictions
+from apisports_football.Models.Sidelined.Sidelined import Sidelined
+from apisports_football.Models.Standings.Standings import Standings
+from apisports_football.Models.Teams.Statistics import Statistics as TeamsStatistics
+from apisports_football.Models.Teams.Teams import Teams
+from apisports_football.Models.Timezone.Timezone import Timezone
+from apisports_football.Models.Transfers.Transfers import Transfers
+from apisports_football.Models.Trophies.Trophies import Trophies
+from apisports_football.Models.Venues.Venues import Venues
 
 API_RESPONSES_PATH = f'{os.getcwd()}/tests/data/api_responses'
 
@@ -52,7 +63,7 @@ class TestPydanticModels:
     def test_fixtures_statistics(self):
         with open(f'{API_RESPONSES_PATH}/fixtures_statistics.json') as file:
             json_data = json.load(file)
-            Statistics.model_validate(json_data)
+            FixturesStatistics.model_validate(json_data)
 
     def test_fixtures_events(self):
         with open(f'{API_RESPONSES_PATH}/fixtures_events.json') as file:
@@ -102,7 +113,7 @@ class TestPydanticModels:
     def test_odds_live(self):
         with open(f'{API_RESPONSES_PATH}/odds_live.json') as file:
             json_data = json.load(file)
-            ...
+            Live.model_validate(json_data)
 
     def test_odds_live_bets(self):
         with open(f'{API_RESPONSES_PATH}/odds_live_bets.json') as file:
@@ -123,8 +134,93 @@ class TestPydanticModels:
         with open(f'{API_RESPONSES_PATH}/players_profiles.json') as file:
             json_data = json.load(file)
             Profiles.model_validate(json_data)
+    
+    def test_players_seasons(self):
+        with open(f'{API_RESPONSES_PATH}/players_seasons.json') as file:
+            json_data = json.load(file)
+            Seasons.model_validate(json_data)
+    
+    def test_players_squads(self):
+        with open(f'{API_RESPONSES_PATH}/players_squads.json') as file:
+            json_data = json.load(file)
+            Squads.model_validate(json_data)
 
     def test_players_teams(self):
         with open(f'{API_RESPONSES_PATH}/players_teams.json') as file:
             json_data = json.load(file)
+            PlayersTeams.model_validate(json_data)
+    
+    def test_players_topassists(self):
+        with open(f'{API_RESPONSES_PATH}/players_topassists.json') as file:
+            json_data = json.load(file)
+            Players.model_validate(json_data)
+    
+    def test_players_topredcards(self):
+        with open(f'{API_RESPONSES_PATH}/players_topredcards.json') as file:
+            json_data = json.load(file)
+            Players.model_validate(json_data)
+    
+    def test_players_topscorers(self):
+        with open(f'{API_RESPONSES_PATH}/players_topscorers.json') as file:
+            json_data = json.load(file)
+            Players.model_validate(json_data)
+    
+    def test_players_topyellowcards(self):
+        with open(f'{API_RESPONSES_PATH}/players_topyellowcards.json') as file:
+            json_data = json.load(file)
+            Players.model_validate(json_data)
+    
+    def test_predictions(self):
+        with open(f'{API_RESPONSES_PATH}/predictions.json') as file:
+            json_data = json.load(file)
+            Predictions.model_validate(json_data)
+
+    def test_sidelined(self):
+        with open(f'{API_RESPONSES_PATH}/sidelined.json') as file:
+            json_data = json.load(file)
+            Sidelined.model_validate(json_data)
+
+    def test_standings(self):
+        with open(f'{API_RESPONSES_PATH}/standings.json') as file:
+            json_data = json.load(file)
+            Standings.model_validate(json_data)
+
+    def test_teams(self):
+        with open(f'{API_RESPONSES_PATH}/teams.json') as file:
+            json_data = json.load(file)
             Teams.model_validate(json_data)
+    
+    def test_teams_countries(self):
+        with open(f'{API_RESPONSES_PATH}/teams_countries.json') as file:
+            json_data = json.load(file)
+            Countries.model_validate(json_data)
+    
+    def test_teams_seasons(self):
+        with open(f'{API_RESPONSES_PATH}/teams_seasons.json') as file:
+            json_data = json.load(file)
+            Seasons.model_validate(json_data)
+
+    def test_teams_statistics(self):
+        with open(f'{API_RESPONSES_PATH}/teams_statistics.json') as file:
+            json_data = json.load(file)
+            TeamsStatistics.model_validate(json_data)
+
+    def test_timezone(self):
+        with open(f'{API_RESPONSES_PATH}/timezone.json') as file:
+            json_data = json.load(file)
+            Timezone.model_validate(json_data)
+
+    def test_transfers(self):
+        with open(f'{API_RESPONSES_PATH}/transfers.json') as file:
+            json_data = json.load(file)
+            Transfers.model_validate(json_data)
+    
+    def test_trophies(self):
+        with open(f'{API_RESPONSES_PATH}/trophies.json') as file:
+            json_data = json.load(file)
+            Trophies.model_validate(json_data)
+
+    def test_venues(self):
+        with open(f'{API_RESPONSES_PATH}/venues.json') as file:
+            json_data = json.load(file)
+            Venues.model_validate(json_data)